Test of Branching + Transitions
-- the old stuff is there but disabled until the transitions are coming along
for some reasons they aren't populating, and fuck all if I know why...
Going to go "take a bath" and see what I can do when I return.

maybe there is an issue with the dayTransDay thing (perhaps not being created)?
there is also some disconcering with the for loop, in that it's only hitting the vibe once, as opposed to the 4 times it was supposed to).