Stopping Point
Just need to figure out the closing animation, 
prob. fix the watchface (i know that'll fuck up somehow)
get the word transition (which should be a challenge on its own)
and then configure the HTTP stuff (and have it properly convert the weather)

maaaaaaaaaaaaaaaayyyyybe do moon stuff a-la persona 3 for the moon phases... @ night on clear nights only?