Intro 90 percent done
Able to trigger @ start of every day (and return to watchface)
- need to flip intro weather from debug to live.
Just need to add logic to switch from one month to the next + one year to the next. 
base framework is in place, just need to code it.

Need to add in word intros
Add in TV intro, and then this will be DONE
- Not sure of persistent storage, since it would still have to communicate with the phone, and therefore not be able to "instantly" get it, which is what I wanted.

May create ability to "bypass" TV intro (search for weather or timeout) by pressing the backlight button, which would go straight to watchface.